==> notebooks/1.0-ys-EDA.ipynb <==
from pathlib import Path

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
sns.set()

from pandas_profiling import ProfileReport

DATA_PATH = Path('../data/raw/')

# Employee

employee_df = pd.read_csv(DATA_PATH / 'sotrudniki.csv', sep=';', parse_dates=['date'])

employee_df.head()

employee_df.info()

profile = ProfileReport(employee_df, title='Employees dataframe')
profile.to_widgets()

# Relatives

relatives_df = pd.read_csv(DATA_PATH / 'rodstvenniki.csv', sep=';')

relatives_df.head()

relatives_df.info()

profile = ProfileReport(relatives_df, title='Relatives dataframe')
profile.to_widgets()

# OGRV

ogrv_df = pd.read_csv(DATA_PATH / 'OGRV.csv', sep=';', parse_dates=['date'])

ogrv_df.head()

ogrv_df.info()

profile = ProfileReport(ogrv_df, title='OGRV dataframe')
profile.to_widgets()
