==> notebooks/1.4-blending.ipynb <==
import numpy as np
import pandas as pd

from pathlib import Path

subpaths = sorted(Path('../data/submissions/blend/').glob('*csv'))

subs = [pd.read_csv(path, sep=';') for path in subpaths]

sub_blend = subs[0].drop('target', axis=1).copy()

for idx, sub in enumerate(subs):
    sub_blend[f'target_{idx}'] = sub['target']

sub_blend['target'] = (
    0.25299710624224886 * sub_blend['target_0'] + 
    0.26027397260273977 * sub_blend['target_1'] + 
    0.25696378830083566 * sub_blend['target_2'] + 
    0.2609880749574106 * sub_blend['target_3'] + 
    0.25811965811965815 * sub_blend['target_4'] + 
    0.2537719515211476 * sub_blend['target_5']) / (0.25299710624224886 + 0.26027397260273977 + 0.25696378830083566 + 0.2609880749574106 + 0.25811965811965815 + 0.2537719515211476)

sub_blend['target'] = sub_blend['target'].astype('int')

sub_blend = sub_blend.drop(['target_0', 'target_1', 'target_2', 'target_3', 'target_4', 'target_5'], axis=1)

sub_blend.to_csv('../data/submissions/blend2.csv', sep=';', index=False)
