==> notebooks/04_dim-validation_scheme.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split
from sklearn.metrics import precision_recall_curve, roc_auc_score, f1_score
from tqdm import tqdm
import json

pd.set_option('display.max_columns', 1000)
pd.set_option('display.max_rows', 1000)
pd.options.mode.chained_assignment = None
np.seterr(divide='ignore', invalid='ignore')

from datetime import datetime, timedelta
from dateutil.relativedelta import relativedelta

def to_datetime(date):
    """
    Converts a numpy datetime64 object to a python datetime object 
    Input:
      date - a np.datetime64 object
    Output:
      DATE - a python datetime object
    """
    timestamp = ((date - np.datetime64('1970-01-01T00:00:00')) / np.timedelta64(1, 's'))
    return datetime.utcfromtimestamp(timestamp)

# Считывание данных

sot = pd.read_csv('sotrudniki.csv', sep = ';')
sot['date'] = pd.to_datetime(sot['date'], format='%Y-%m-%d')

train_target_df = sot[['hash_tab_num', 'date', 'sick']]
train_target_df.head()

---

sot_data = sot[[
    'hash_tab_num','date','category','gender','razryad_fact','work_experience_company',
    'name_fact_lvl5','education','home_to_work_distance'
]]

sot_data['gender'] = sot_data['gender'].map(lambda x: 1 if x == 'мужской' else 0)

# Создание вспомогательно датасета с информацией о количестве сотрудников в подразделении
# по фактическому месту работы

division_count = sot_data[['hash_tab_num','date','name_fact_lvl5']].\
groupby(['name_fact_lvl5','date']).agg('count').reset_index()

division_count.columns = ['name_fact_lvl5', 'date', 'personel_num']

sot_data = pd.merge(sot_data, division_count, how = 'left', on = ['date','name_fact_lvl5'])

# Создание dummy переменных

sot_data.education = sot_data['education']\
.map(lambda x: 'Высшее' if x in ['Высшее образование','Высшее-бакалавриат','Высшее-специалитет'] else(\
'Среднее_профессинальное' if x in ['Ср.профессиональное','Нач.профессиональное'] else 'Начальное_среднее'))

sot_data = pd.get_dummies(sot_data, columns = ['category','education','razryad_fact']).drop('name_fact_lvl5', axis = 1)
sot_data['orig_date'] = sot_data['date'].copy()

sot_data.head()

submission_extra = sot_data[sot_data['orig_date'] == pd.to_datetime('2019-08-01')]
submission_extra['target'] = 0

---

with open('transformed_data/date_of_birth.json', 'r') as f:
    date_of_birth_dict = json.load(f)
    date_of_birth_dict = {int(k): int(v) for k, v in date_of_birth_dict.items()}


def calc_age(hash_tab_num, calc_date, date_of_birth_dict):
    cur_date = int(calc_date)
    birth_date = date_of_birth_dict[hash_tab_num]
    age = cur_date - birth_date
    return age


with open('transformed_data/relatives_info.json', 'r') as f:
    relatives_dict = json.load(f)
    relatives_dict = {int(k): v for k, v in relatives_dict.items()}


def calc_relatives_bins(hash_tab_num, calc_date, relatives_dict):
    '''
    bins:
        0: 0 - 3: младенец
        1: 4 - 7: ребенок
        2: 8 - 18: школьник
        3: 19 - 35: молодежь :)
        4: 36 - 55(F), 60(M): предпенсионный возраст
        5: 55(F), 60(M) - +++: пенсионер
        6: кол-во родственников мужского рода
        7: кол-во родственников женского рода
    '''
    
    bins = [0] * 8
    if hash_tab_num not in relatives_dict:
        return bins
    
    cur_date = int(calc_date)
    for (sex, birth_date) in relatives_dict[hash_tab_num]:
        if sex == 'M':
            bins[6] += 1
        elif sex == 'F':
            bins[7] += 1
            
        if birth_date < 0:
            continue
            
        age = cur_date - birth_date
        if age < 0:
            continue
        elif age <= 3:
            bins[0] += 1
        elif age <= 7:
            bins[1] += 1
        elif age <= 18:
            bins[2] += 1
        elif age <= 35:
            bins[3] += 1
        else:
            if (sex == 'M' and age >= 60) or (sex == 'F' and age >= 55):
                bins[5] += 1
            else:
                bins[4] += 1
    return bins

def target_date_features(df):
    df['year'] = df['date'].dt.year
    df['age'] = df.apply(lambda x: calc_age(x['hash_tab_num'], x['year'], date_of_birth_dict), axis=1)
    df['is_pensioner'] = (((df['age'] >= 60) & (df['gender'] == 1)) | ((df['age'] >= 55) & (df['gender'] == 0))).astype(int)
    df['relatives'] = df.apply(lambda x: calc_relatives_bins(x['hash_tab_num'], x['year'], relatives_dict), axis=1)
    for i in range(8):
        df[f'relatives_{i}'] = df['relatives'].apply(lambda x: x[i])
    df = df.drop(columns=['year', 'relatives'])
    
    return df

## 1 Month model

def run_training(months, check_shift):
    prediction_dates = sorted(sot_data['orig_date'].unique())[:-(months + check_shift)]
    
    roc_auc_scores = []
    thresholds = []
    f1_scores = []
    last_model = None

    for idx, prediction_date in enumerate(prediction_dates):

        df_train = sot_data[sot_data['orig_date'] <= prediction_date].copy()
        df_train['date'] = df_train['orig_date'] + pd.DateOffset(months=months)
        df_train = pd.merge(df_train, train_target_df, on=['hash_tab_num', 'date'])

        test_date = to_datetime(prediction_date) + relativedelta(months=check_shift)
        df_test = sot_data[sot_data['orig_date'] == test_date].copy()
        df_test['date'] = df_test['orig_date'] + pd.DateOffset(months=months)
        df_test = pd.merge(df_test, train_target_df, on=['hash_tab_num', 'date'])

        df_train = target_date_features(df_train)
        df_test = target_date_features(df_test)

        X_train = df_train.drop(columns=['hash_tab_num', 'date', 'orig_date', 'sick']).fillna(-100)
        X_test = df_test.drop(columns=['hash_tab_num', 'date', 'orig_date', 'sick']).fillna(-100)

        y_train = df_train['sick']
        y_test = df_test['sick']

        model = RandomForestClassifier()
        model.fit(X_train, y_train)
        preds = model.predict_proba(X_test)[:, 1]
        roc_auc_score_ = roc_auc_score(y_test, preds)
        roc_auc_scores.append(roc_auc_score_)

        if len(thresholds) > 0:
            th = np.mean(thresholds)
            f1_score_ = f1_score(y_test, (preds > th).astype(int))
            f1_scores.append(f1_score_)
        else:
            f1_score_ = 0


        p, r, thresholds_ = precision_recall_curve(y_test, preds)
        f1_scores_ = 2 * r * p / (r + p)
        f1_scores_ = f1_scores_[p > 0]
        th = thresholds_[np.argmax(f1_scores_)]
        thresholds.append(th)

        last_model = model
        print(f'{idx + 1}/{len(prediction_dates)}', roc_auc_score_, f1_score_, th, sep='\t')
        
    return last_model, roc_auc_scores, thresholds, f1_scores

m1 = run_training(months=1, check_shift=1)

m2 = run_training(months=2, check_shift=1)

m3 = run_training(months=3, check_shift=1)

m4 = run_training(months=4, check_shift=1)

m5 = run_training(months=5, check_shift=1)

m6 = run_training(months=6, check_shift=1)

m7 = run_training(months=7, check_shift=1)

m8 = run_training(months=8, check_shift=1)

m9 = run_training(months=9, check_shift=1)

m10 = run_training(months=10, check_shift=1)

m11 = run_training(months=11, check_shift=1)

m12 = run_training(months=12, check_shift=1)

result_models = [m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12]

---

results = []

for month in range(1, 13):
    submission_extra['date'] = submission_extra['orig_date'] + pd.DateOffset(months=month)
    new_df = target_date_features(submission_extra)
    X = new_df.drop(columns=['hash_tab_num', 'date', 'orig_date', 'target']).fillna(-100)
    
    model = result_models[month - 1][0]
    th = np.mean(result_models[month - 1][2])
    submission_extra['target'] = (model.predict_proba(X)[:, 1] >= th).astype(int)
    results.append(submission_extra[['hash_tab_num', 'date', 'target']].copy())
    
result_df = pd.concat(results, ignore_index=True)
result_df.head()

check_df = pd.read_csv('submission_check.csv', sep=';').drop(columns=['target'])
check_df['date'] = pd.to_datetime(check_df['date'], format='%Y-%m-%d')
check_df.head()

result_df_new = pd.merge(result_df, check_df, on=['hash_tab_num', 'date']).sort_values(['hash_tab_num', 'date'])
result_df_new.head()

result_df_new.to_csv('submission_8.csv', sep=';', index=False)

X = new_df.drop(columns=['hash_tab_num', 'date', 'orig_date']).fillna(-100)

X_train.columns

np.mean(m2[2]), np.mean(m2[2][-10:])

target_date_features(submission_extra)

submission_extra
